a2a/pyproject.toml
[project]
name = "a2a-examples"
version = "0.1.0"
description = "A2A tutorial course working examples"
requires-python = ">=3.10"
license = { text = "MPL-2.0" }

dependencies = [
  "openai>=1.30.0",
  "a2a-sdk[http-server]>=0.3.24",
  "uvicorn[standard]>=0.27.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "jupyter>=1.0.0",
  "ipykernel>=6.0.0",
  "nbconvert>=7.0.0",
  # Lesson 08 — Microsoft Agent Framework
  "agent-framework>=1.0.0rc2",
  "agent-framework-azure-ai>=1.0.0rc2",
  "pydantic>=2.0.0",
  # Lesson 09 — Google Agent Development Kit
  "google-adk[a2a]>=1.3.0",
  "litellm>=1.50.0",
]

[tool.uv]
dev-dependencies = ["ruff>=0.15.4"]
